Refactor ECS components and systems for improved clarity
- Removed logging setup and ECS initialization from App.xaml.cs for a streamlined startup.
- Updated EnergyComponent and HealthComponent to use float types for more precise values.
- Added PositionComponent to represent entity positions with X and Y properties.
- Enhanced MovementComponent with DX and DY properties for position changes.
- Refactored EnergySystem and HealthSystem to eliminate logger dependencies, using console output instead.
- Introduced MovementSystem to manage entity movement with random logic.
- Updated EnergyUtility and HealthUtility to handle float amounts and removed logger parameters.
- Modified MainWindow.xaml to change the title and add an Image control for rendering.
- Expanded MainWindow.xaml.cs to include a timer for ECS updates and entity rendering.
- Updated project file to include Microsoft.Extensions.Logging.Abstractions package reference.

+++ EvoSim/ECS/Components/EnergyComponent.cs
@@ -4,6 +4,6 @@ namespace EvoSim.ECS.Components;
 
 public class EnergyComponent : IComponent
 {
-    public int Energy;
-    public int MaxEnergy;
+    public float Energy;
+    public float MaxEnergy;
 }
