+++ EvoSim/ECS/Components/HealthComponent.cs
@@ -4,7 +4,7 @@ namespace EvoSim.ECS.Components;
 
 public class HealthComponent : IComponent
 {
-    public int Health;
-    public int MaxHealth;
+    public float Health;
+    public float MaxHealth;
     public bool IsAlive => Health > 0;
 }
